// src/dates/convert-units/time-unit.enum.ts
export enum TimeUnit {
  Millisecond = 'millisecond',
  Milliseconds = 'milliseconds',
  Second = 'second',
  Seconds = 'seconds',
  Minute = 'minute',
  Minutes = 'minutes',
  Hour = 'hour',
  Hours = 'hours',
  Day = 'day',
  Days = 'days',
  Week = 'week',
  Weeks = 'weeks',
  Month = 'month',
  Months = 'months',
  Year = 'year',
  Years = 'years',
}